add follow teacher scenario fix models add pull to refresh in follows screen

+++ drosak/lib/lectures/model/entity/lecture.dart
@@ -14,7 +14,7 @@ class Lecture {
   String teacherName;
   String teacherImageUrl;
   DateTime? bookingDate;
-  int? teacherRating;
+  double? teacherRating;
 
   Lecture(
       {required this.centerName,
